+++ Assets/Scripts/Structs/WeaponData.cs
@@ -15,7 +15,7 @@ public struct WeaponData
     public float atkrange;
     public int maxaim;
     public int maxcross;
-    public float repel;
+    public int repel;
     public float critical;
 
     public void Init()
